backend/requirements.txt
# MedVoice AI - Python Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-dotenv==1.0.0
httpx==0.26.0

# Voice Processing
deepgram-sdk==3.0.0
google-cloud-texttospeech==2.16.0

# LLM (OpenRouter - OpenAI compatible)
openai==1.12.0

# Firebase
firebase-admin==6.4.0
google-cloud-firestore==2.14.0

# Twilio
twilio==8.13.0

# Data Models
pydantic==2.6.0
pydantic-settings==2.1.0

# Utilities
python-multipart==0.0.6
audioop-lts==0.2.1  # Audio processing for mulaw conversion

# Development
pytest==8.0.0
pytest-asyncio==0.23.0
